requirements.txt
aiohttp==3.9.4
aiohttp-retry==2.8.3
aiosignal==1.3.1
attrs==23.2.0
backports.tarfile==1.0.0
beautifulsoup4==4.12.3
certifi==2024.2.2
charset-normalizer==3.3.2
frozenlist==1.4.1
idna==3.7
jaraco.context==5.3.0
more-itertools==10.2.0
multidict==6.0.5
PyAudio==0.2.14
PyJWT==2.8.0
pyttsx3==2.90
requests==2.31.0
soupsieve==2.5
SpeechRecognition==3.10.3
twilio==9.0.4
typing_extensions==4.11.0
urllib3==2.2.1
wikipedia==1.4.0
wolframalpha==5.0.0
xmltodict==0.13.0
yarl==1.9.4
